+++ src/pages/fans/my-tickets/my-tickets-tabs/active-tab.tsx
@@ -15,26 +15,27 @@ export default function ActiveTicketsTab({
   }
 
   return (
-    <>
+    <div className='w-full flex flex-col items-center justify-center'>
       {isEmpty ? (
         <EmptyState type='active' btn_name='Discover Events' path={getRoutePath('events')} />
       ) : (
         <ActiveTickets data={data} />
       )}
-    </>
+    </div>
   )
 }
 
 function ActiveTickets({ data }: { data: UserTicketData[] }) {
   return (
-    <div className='flex flex-wrap overflow-y-auto items-center justify-center gap-7 px-5 md:px-[50px] lg:px-[100px] mb-[100px]'>
+    <div className='grid w-full h-screen grid-cols-2 md:grid-cols-4 lg:grid-cols-6 '>
       {data.map((item) => (
         <Tickets
           key={item.eventId}
           id={item.eventId}
           event_name={item.eventName}
           image={item.desktopMedia?.flyer}
-          quantity={item.quantity}
+          quantity={item.ticketDetails.reduce((sum, t) => sum + t.totalQuantity, 0)}
+          ticketName={item.ticketDetails[0]?.ticketName ?? ''}
         />
       ))}
     </div>