+++ src/pages/landing-page/resale-marketplace/individual-resale-event/index.tsx
@@ -1,24 +1,24 @@
-import { SEO } from "@/components/seo";
-import { events } from "@/data/events";
-import { useParams } from "react-router-dom";
-import ResaleEventDetails from "./resale-event-details";
+// import { SEO } from "@/components/seo";
+// import { events } from "@/data/events";
+// import { useParams } from "react-router-dom";
+// import ResaleEventDetails from "./resale-event-details";
 
-export default function IndividualResaleTicket() {
-  const { eventId } = useParams();
-  const event = events.find((item) => item.id === Number(eventId));
+// export default function IndividualResaleTicket() {
+//   const { eventId } = useParams();
+//   const event = events.find((item) => item.id === Number(eventId));
 
-  if (!event) {
-    return <p>No Events Found</p>;
-  }
+//   if (!event) {
+//     return <p>No Events Found</p>;
+//   }
 
-  return (
-    <>
-      <SEO
-        title={`${event.event_name} - Afro Revive`}
-        description={`Buy tickets for ${event.event_name} - Afro Revive`}
-      />
+//   return (
+//     <>
+//       <SEO
+//         title={`${event.event_name} - Afro Revive`}
+//         description={`Buy tickets for ${event.event_name} - Afro Revive`}
+//       />
 
-      <ResaleEventDetails event={event} />
-    </>
-  );
-}
+//       <ResaleEventDetails event={event} />
+//     </>
+//   );
+// }