+++ src/types/profile.ts
@@ -10,7 +10,6 @@ export interface UpdateUserProfileRequest {
   dateOfBirth: string
   country: string
   state: string
-  password: string
 }
 
 export interface CompleteProfileRequest {
@@ -51,17 +50,11 @@ export type UserProfileResponse = ApiResponse<UserProfileData>
 
 // User Ticket Data for GET /api/Profile/user/ticket/active and /api/Profile/user/ticket/past
 export interface UserTicketData {
-  ticketId: string
-  purchaseDate: string
-  status: string
   eventId: string
   eventName: string
   eventVenue: string
   eventStartDate: string
   eventEndDate: string
-  ticketName: string
-  price: number
-  quantity: number
   theme: {
     themeName: string
   }
@@ -69,6 +62,19 @@ export interface UserTicketData {
     flyer: string
     background: string
   }
+  ticketDetails: UserTicketTicketDetails[]
+}
+
+export interface UserTicketTicketDetails {
+  ticketId: string
+  ticketName: string
+  price: number
+  totalQuantity: number
+  purchaseHistory: {
+    orderId: string
+    purchaseDate: string
+    quantity: number
+  }[]
 }
 
 export type UserTicketsResponse = ApiResponse<UserTicketData>